refactor: update component repositories and files to PascalCase
