+++ src/components/MovieCard/MovieCard.scss
