+++ src/components/MovieView/MovieView.scss
